Fixed dropdown crashing the page
